refactor: move CLI sources into src/cli/
