+++ package.json
@@ -6,7 +6,7 @@
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "bin": {
-    "openapi-ts-json-schema": "./dist/cli.js"
+    "openapi-ts-json-schema": "./dist/cli/cli.js"
   },
   "files": [
     "dist"
