+++ test/cli-entry.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from 'vitest';
 
-import { cliCommand } from '../src/cliCommand.js';
+import { cliCommand } from '../src/cli/cliCommand.js';
 
 const { runMain } = vi.hoisted(() => ({ runMain: vi.fn() }));
 
@@ -11,7 +11,7 @@ vi.mock('citty', async (importOriginal) => {
 
 describe.skip('cli entry point', () => {
   it('invokes runMain with cliCommand on startup', async () => {
-    await import('../src/cli.js');
+    await import('../src/cli/cli.js');
 
     expect(runMain).toHaveBeenCalledTimes(1);
     expect(runMain).toHaveBeenCalledWith(cliCommand);
