+++ tsconfig.build.json
@@ -1,6 +1,6 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["./src/index.ts", "./src/cli.ts"],
+  "include": ["./src/index.ts", "./src/cli/cli.ts"],
   "compilerOptions": {
     "outDir": "dist",
     "declaration": true
